requirements.txt
# Core libraries
faiss-cpu==1.7.4
PyMuPDF==1.23.6
sentence-transformers==2.2.2
google-generativeai==0.4.1

# TensorFlow and Keras (used indirectly via tf_keras)
tensorflow==2.13.0
tf-keras==0.5.0

# Optional: suppress logging noise (if needed)
absl-py==1.4.0
